[Fix] 신장 / 체중 Double 형 -> Int형으로 수정

+++ data/src/main/java/com/dkproject/data/model/UserDTO.kt
@@ -7,8 +7,8 @@ data class UserDTO(
     @DocumentId val id: String,
     val nickName: String,
     val position: List<String>,
-    val weight: Double?,
-    val height: Double?,
+    val weight: Int?,
+    val height: Int?,
     val profileImageUrl: String
 )
